complete 26

+++ Hazel/src/Hazel/Events/MouseEvent.h
@@ -51,7 +51,7 @@ namespace Hazel
 	class HAZEL_API MouseButtonEvent :public Event
 	{
 	public:
-		inline float GetMouseButton() const { return m_Button; }
+		inline int GetMouseButton() const { return m_Button; }
 
 		EVENT_CLASS_CATEGORY(EventMouseButton | EventCategoryInput)
 	protected:
